ml-models/requirements.txt
tensorflow==2.0.0b0
google-cloud-storage==1.16.1
scikit-learn==0.22.2
requests==2.24.0